Set browser location in an attempt to gracefully handle browser reload
The issue is that the EO1 seems to be reloading the page after somewhere between 50 and 250 frames, and because the initial frame is in the URL, it goes back to where it started. I'd rather not have to fix the current frame manually every time, so this is the last thing I can think to try to fix it in the browser. We'll see if, when it runs out of memory again, it reloads the current URL, or reloads what it received in `set_url` in the device OS. If this doesn't work, I'm going to need to keep the current state on the web server :-/

I learned about the URLSearchParams and window.history.replace state here: https://gist.github.com/excalq/2961415#gistcomment-2221360